Split out special task queue
